GCP/Cloud Function/requirements.txt
# Function dependencies, for example:
# package>=version

json5
google-cloud-storage
numpy
pandas

